requirements.txt
PyYAML
requests
tabulate
urllib3
json-diff
jsondiff
datetime
